default.html: Add development documentation and fix a typo
Until the project fully migrates to ReadTheDocs in the next
stable release, the documentation link on the homepage is
constantly outdated.

This commit add the ReadTheDocs link to the project homepage,
as "latest documentation" (the original documentation is
renamed to "stable documentation").

This commit also fixes the typo of "GitHub" as "Gihub".

Signed-off-by: Yifeng Li <tomli@tomli.me>